+++ TagCloudContainer/IUserInterface.cs
@@ -1,0 +1,7 @@
+﻿namespace TagCloudContainer
+{
+    public interface IUserInterface
+    {
+        Config GetConfig(string[] args, out bool toExit);
+    }
+}
