--- TagCloudLayouter/IUserInterface.cs
@@ -1,7 +1,0 @@
-﻿namespace TagCloudLayouter
-{
-    public interface IUserInterface
-    {
-        Config GetConfig(string[] args, out bool toExit);
-    }
-}
