--- TagsCloudContainer/IPreprocessor.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace TagsCloudContainer
-{
-    public interface IPreprocessor
-    {
-        IEnumerable<string> GetValidWords(IEnumerable<string> words);
-    }
-}
