--- TagsCloudContainer/IReader.cs
@@ -1,7 +1,0 @@
-﻿namespace TagsCloudContainer
-{
-    public interface IReader
-    {
-        string ReadFromFile(string path);
-    }
-}
