--- TagsCloudContainer/ITextParser.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace TagsCloudContainer
-{
-    public interface ITextParser
-    {
-        IEnumerable<string> GetWords(string text);
-    }
-}
