--- TagsCloudContainer/IWordExcluder.cs
@@ -1,10 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace TagsCloudContainer
-{
-    public interface IWordExcluder
-    {
-        HashSet<string> GetExcludedWords();
-        void SetExcludedWord(string word);
-    }
-}
