--- TagsCloudContainer/TxtReader.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.IO;
-
-namespace TagsCloudContainer
-{
-    public class TxtReader:IReader
-    {
-        public string ReadFromFile(string path)
-        {
-            using (var sr = new StreamReader(path))
-            {
-                return sr.ReadToEnd();
-            }
-        }
-    }
-}
