+++ TagsCloudPreprocessor/IPreprocessor.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace TagsCloudPreprocessor
+{
+    public interface IPreprocessor
+    {
+        IEnumerable<string> GetValidWords(IEnumerable<string> words);
+    }
+}
