+++ TagsCloudPreprocessor/IReader.cs
@@ -1,0 +1,7 @@
+﻿namespace TagsCloudPreprocessor
+{
+    public interface IReader
+    {
+        string ReadFromFile(string path);
+    }
+}
