+++ TagsCloudPreprocessor/ITextParser.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace TagsCloudPreprocessor
+{
+    public interface ITextParser
+    {
+        IEnumerable<string> GetWords(string text);
+    }
+}
