+++ TagsCloudPreprocessor/IWordExcluder.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace TagsCloudPreprocessor
+{
+    public interface IWordExcluder
+    {
+        HashSet<string> GetExcludedWords();
+        void SetExcludedWord(string word);
+    }
+}
